remove storybook move tooltip component
