+++ src/Common/Tooltip/index.tsx
@@ -1,7 +1,7 @@
 import { useId } from "@uifabric/react-hooks";
 import { ITooltipHostStyles, TooltipHost } from "office-ui-fabric-react/lib/Tooltip";
 import * as React from "react";
-import InfoBubble from "../../../../../images/info-bubble.svg";
+import InfoBubble from "../../../images/info-bubble.svg";
 
 const calloutProps = { gapSpace: 0 };
 const hostStyles: Partial<ITooltipHostStyles> = { root: { display: "inline-block" } };
@@ -12,11 +12,13 @@ export interface TooltipProps {
 export const Tooltip: React.FunctionComponent = ({ children }: TooltipProps) => {
   const tooltipId = useId("tooltip");
 
-  return (
+  return children ? (
     <span>
       <TooltipHost content={children} id={tooltipId} calloutProps={calloutProps} styles={hostStyles}>
         <img className="infoImg" src={InfoBubble} alt="More information" />
       </TooltipHost>
     </span>
+  ) : (
+    <></>
   );
 };
